lifecycle working

+++ tealium/dispatchqueue/TealiumPersistentDispatchQueue.swift
@@ -14,59 +14,38 @@ import TealiumCore
 class TealiumPersistentDispatchQueue {
 
     private var diskStorage: TealiumDiskStorageProtocol!
-//    var queuedDispatches = [
+    public var currentEvents: Int = 0
 
     public init(diskStorage: TealiumDiskStorageProtocol) {
         self.diskStorage = diskStorage
-//        initializeQueue()
+        if let totalEvents = self.peek()?.count {
+            self.currentEvents = totalEvents
+        }
+
     }
 
-//    func initializeQueue() {
-//        diskStorage.retrieve(as: [TealiumTrackRequest].self) { _, data, _ in
-//            guard let data = data else {
-//
-//                return
-//            }
-////            self.setLoadedAppData(data: data)
-//        }
-////        // queue already initialized
-////        if let _ = TealiumPersistentDispatchQueue.queueStorage.object(forKey: storageKey) as? [[String: Any]] {
-////            return
-////        }
-////        // init with blank data
-////        let blankData = [[String: Any]]()
-////        TealiumPersistentDispatchQueue.queueStorage.set(blankData, forKey: storageKey)
-//    }
-
     func saveDispatch(_ dispatch: TealiumTrackRequest) {
+        // TODO: do something useful with completion
+        currentEvents += 1
         diskStorage.append(dispatch, completion: nil)
     }
 
+    func saveAndOverwrite(_ dispatches: [TealiumTrackRequest]) {
+        // TODO: do something useful with completion
+        currentEvents = dispatches.count
+        diskStorage.save(dispatches, completion: nil)
+    }
+
     func peek() -> [TealiumTrackRequest]? {
-//        return dequeueDispatches(clear: false)
-        var trackRequests: [TealiumTrackRequest]?
-        diskStorage.retrieve(as: [TealiumTrackRequest].self) { _, data, error in
-            guard error == nil else {
-                trackRequests = nil
-                return
-            }
-            trackRequests = data
+        guard let dispatches = dequeueDispatches(clear: false) else {
+            return nil
         }
-        return trackRequests
+        currentEvents = dispatches.count
+        return dispatches
     }
 
     func dequeueDispatches(clear clearQueue: Bool? = true) -> [TealiumTrackRequest]? {
         var queuedDispatches: [TealiumTrackRequest]?
-//        readWriteQueue.read {
-//            if let dispatches = TealiumPersistentDispatchQueue.queueStorage.array(forKey: self.storageKey) as? [[String: Any]] {
-//                // clear persistent queue
-//                if clearQueue == true {
-//                    self.clearQueue()
-//                }
-//                queuedDispatches = dispatches
-//            }
-//        }
-//        return queuedDispatches
         diskStorage.retrieve(as: [TealiumTrackRequest].self) { _, data, error in
             guard error == nil else {
                 queuedDispatches = nil
@@ -75,7 +54,10 @@ class TealiumPersistentDispatchQueue {
             queuedDispatches = data
         }
 
-        diskStorage.delete(completion: nil)
+        if clearQueue == true {
+            self.currentEvents = 0
+            diskStorage.delete(completion: nil)
+        }
 
         return queuedDispatches
     }
@@ -86,7 +68,7 @@ class TealiumPersistentDispatchQueue {
             return
         }
 
-        // note: any completion blocks will be ignored for now, since we can only persist Dictionaries in UserDefaults
+        // note: any completion blocks will be ignored for now, since we can only persist Dictionaries
         var newData = currentData
         let totalDispatches = newData.count
         if totalDispatches == maxQueueSize {
@@ -95,17 +77,13 @@ class TealiumPersistentDispatchQueue {
             // left with 19 elements => 20 - (20-1) = 19
             let slice = newData.suffix(from: totalDispatches - (maxQueueSize - 1))
             newData = Array(slice)
-//            readWriteQueue.write {
-//                TealiumPersistentDispatchQueue.queueStorage.set(newData, forKey: self.storageKey)
-//            }
+            self.saveAndOverwrite(newData)
         }
     }
 
     func clearQueue() {
-//        readWriteQueue.write {
-//            let blankData = [[String: Any]]()
-//            TealiumPersistentDispatchQueue.queueStorage.set(blankData, forKey: self.storageKey)
-//        }
+        // TODO: do something useful with completion
+        diskStorage.delete(completion: nil)
     }
 
 }
